fix: add unitType for ingredient

+++ backend/module/ingredient/ingredientmodel/ingredient_update_amount.go
@@ -3,7 +3,7 @@ package ingredientmodel
 import "backend/common"
 
 type IngredientUpdateAmount struct {
-	Amount int `json:"amount" gorm:"-"`
+	Amount float32 `json:"amount" gorm:"-"`
 }
 
 func (*IngredientUpdateAmount) TableName() string {
