+++ backend/module/stockchangehistory/stockchangehistorymodel/stock_change_history.go
@@ -5,8 +5,8 @@ import "backend/common"
 type StockChangeHistory struct {
 	Id           string           `json:"id" gorm:"column:id;"`
 	IngredientId string           `json:"ingredientId" gorm:"column:ingredientId;"`
-	Amount       int              `json:"amount" gorm:"column:amount;"`
-	AmountLeft   int              `json:"amountLeft" gorm:"column:amountLeft;"`
+	Amount       float32          `json:"amount" gorm:"column:amount;"`
+	AmountLeft   float32          `json:"amountLeft" gorm:"column:amountLeft;"`
 	Type         *StockChangeType `json:"type" gorm:"column:type;"`
 }
